openai/prompty_template/requirements.txt
python-dotenv
langchain-core
langchain-openai